api/v1/views/requirement.txt
alembic==1.13.2
aniso8601==9.0.1
bcrypt==4.2.0
blinker==1.8.2
cachelib==0.9.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.1
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-MySQLdb==2.0.0
Flask-RESTful==0.3.10
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
greenlet==3.0.3
gunicorn==23.0.0
idna==3.8
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
msgspec==0.18.6
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
PyMySQL==1.1.1
pytest==8.3.2
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.32
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
